refactor(sync-core): inline single-consumer abstractions, consolidate protocol
- Move encodeSyncRequest, decodeSyncRequest, stateVectorsEqual from storage.ts
  to protocol.ts (wire-format encoding belongs with protocol, not storage)
- Inline handleHttpSync/handleHttpGetDoc into standalone server (sole consumer)
- Inline UpdateLog type into standalone server's storage.ts (sole consumer)
- Inline extractBearerToken into server-local's token-guard.ts (sole consumer)
- Delete auth.ts and storage.ts from sync-core (no remaining consumers)
- Remove compactUpdateLog and createMemoryUpdateLog (unused)

sync-core now exports only what has multiple consumers: protocol encoding,
WS connection lifecycle handlers, and room management. Matches y-sweet's
principle of justified abstractions — every export has 2+ real consumers.
